requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
cffi==1.11.5
colorama==0.4.1
cryptography==2.4.2
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-haystack==2.8.1
django-redis==4.10.0
djangorestframework==3.9.0
drf-extensions==0.4.0
elasticsearch==6.3.1
flower==0.9.2
idna==2.8
isort==4.3.4
kombu==4.2.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
pbr==5.1.1
Pillow==5.3.0
pycodestyle==2.4.0
pycparser==2.19
pylint==2.2.2
PyMySQL==0.9.2
pytz==2018.7
redis==3.0.1
six==1.12.0
stevedore==1.30.0
tornado==5.1.1
typed-ast==1.1.0
urllib3==1.24.1
vine==1.1.4
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
